+++ client/student/router.js
@@ -1,13 +1,13 @@
-import Activity from '@/student/components/program/Activity/index';
+import Activity from '@/student/components/Program/Activity/index';
 import Auth from '@/student/components/auth';
-import Courseware from '@/student/components/program/Courseware/index';
+import Courseware from '@/student/components/Program/Courseware/index';
 import ForgotPassword from '@/student/components/auth/ForgotPassword';
 import get from 'lodash/get';
-import Home from '@/student/components/home/index';
+import Home from '@/student/components/index';
 import Login from '@/student/components/auth/Login';
 import NotFound from '@/admin/components/common/NotFound';
-import Program from '@/student/components/program/index';
-import ProgramSelection from '@/student/components/home/ProgramSelection';
+import Program from '@/student/components/Program/index';
+import ProgramSelection from '@/student/components/ProgramSelection';
 import ResetPassword from '@/student/components/auth/ResetPassword';
 import role from '@/../common/config/role';
 import Router from 'vue-router';
